#365 refactpr : 객체의 책임에 맞게 코드 이동 1. DatesDuration내에서 MonthLineExpense에서만 사용하던 코드 MonthLineResponse객체로 이동 2. 비즈니스 로직을 포함하는 dto -> vo로 도메인 이동
