+++ src/main/java/com/floney/floney/book/service/BookLineService.java
@@ -1,10 +1,10 @@
 package com.floney.floney.book.service;
 
+import com.floney.floney.book.domain.vo.MonthLinesResponse;
 import com.floney.floney.book.dto.process.DayLines;
 import com.floney.floney.book.dto.request.AllOutcomesRequest;
 import com.floney.floney.book.dto.request.BookLineRequest;
 import com.floney.floney.book.dto.response.BookLineResponse;
-import com.floney.floney.book.dto.response.MonthLinesResponse;
 import com.floney.floney.book.dto.response.TotalDayLinesResponse;
 
 import java.util.List;
